requirements.txt
fastapi==0.111.0
uvicorn==0.30.3
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
structlog==24.1.0
redis==5.0.7
rq==1.16.1
opencv-python-headless==4.10.0.84
pymupdf==1.24.9
python-pptx==0.6.23
pytesseract==0.3.13
Pillow==10.4.0
# Manim & ffmpeg are expected to be installed in the runtime image
